requirements.txt
paramiko>=2.11.0
requests>=2.28.0
mysql-connector-python>=8.0.0
fastapi>=0.95.0
uvicorn>=0.22.0
typesense>=0.16.0
websockets>=11.0
python-multipart>=0.0.6
